+++ app/components/AdminCreateAnouncements.js
@@ -1,5 +1,6 @@
 "use client";
 
+import Sidebar from './Sidebar';
 import React, { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import {
@@ -18,102 +19,9 @@ import ExitToAppIcon from '@mui/icons-material/ExitToApp';
 import ReportProblem from '@mui/icons-material/ReportProblem';
 
 export default function AdminCreateAnnouncements() {
-  const [openLogoutDialog, setOpenLogoutDialog] = useState(false);
-  const router = useRouter();
-
-  const handleLogoutClick = () => {
-    setOpenLogoutDialog(true);
-  };
-
-  const handleProfileClick = () => {
-    router.push('/adminProfile');
-  };
-
-  const handleTerminClick = () => {
-    router.push('/adminTermin');
-  };
-
-  const handleLogoutConfirm = () => {
-    setOpenLogoutDialog(false);
-    router.push('/');
-  };
-
-  const handleAnouncements = () => {
-    router.push('/adminAnouncements');
-  };
-
-
-  const handleLogoutCancel = () => {
-    setOpenLogoutDialog(false);
-  };
-
   return (
     <Box sx={{ display: 'flex', height: '100vh' }}>
-      {/* Sidebar */}
-      <Box
-        sx={{
-          width: 250,
-          backgroundColor: '#333',
-          color: '#ccc',
-          paddingTop: 2,
-          display: 'flex',
-          flexDirection: 'column',
-          justifyContent: 'space-between',
-        }}
-      >
-        <List>
-          <ListItemButton>
-            <ListItemIcon>
-              <DashboardIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Dashboard" />
-          </ListItemButton>
-          <ListItemButton>
-            <ListItemIcon>
-              <GroupIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Benutzerverwaltung" />
-          </ListItemButton>
-          <ListItemButton onClick={handleTerminClick}>
-            <ListItemIcon>
-              <EventIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Terminmanagement" />
-          </ListItemButton>
-          <ListItemButton>
-            <ListItemIcon>
-              <NotificationsIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Benachrichtigungen" />
-          </ListItemButton>
-          <ListItemButton onClick={handleAnouncements}>
-            <ListItemIcon>
-              <ReportProblem style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Ankündigungen" />
-          </ListItemButton>
-        </List>
-        <List>
-          <ListItemButton onClick={handleProfileClick}>
-            <ListItemIcon>
-              <AccountCircleIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Profil" />
-          </ListItemButton>
-          <ListItemButton>
-            <ListItemIcon>
-              <SettingsIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Einstellungen" />
-          </ListItemButton>
-          <ListItemButton onClick={handleLogoutClick}>
-            <ListItemIcon>
-              <ExitToAppIcon style={{ color: '#ccc' }} />
-            </ListItemIcon>
-            <ListItemText primary="Logout" />
-          </ListItemButton>
-        </List>
-      </Box>
+      <Sidebar />
 
       {/* Main Content */}
       <Box sx={{ flex: 1, backgroundColor: '#f5f5f5', padding: 3 }}>
@@ -192,27 +100,6 @@ export default function AdminCreateAnnouncements() {
           </Box>
         </Box>
       </Box>
-
-      {/* Logout Confirmation Dialog */}
-      <Dialog
-        open={openLogoutDialog}
-        onClose={handleLogoutCancel}
-      >
-        <DialogTitle>Abmelden</DialogTitle>
-        <DialogContent>
-          <DialogContentText>
-            Möchten Sie sich wirklich abmelden?
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleLogoutCancel} color="primary">
-            Nein
-          </Button>
-          <Button onClick={handleLogoutConfirm} color="primary" autoFocus>
-            Ja
-          </Button>
-        </DialogActions>
-      </Dialog>
     </Box>
   );
 }
